feat(#31):리스트 페이지 구성
