+++ src/types/fancy.d.ts
@@ -10,7 +10,11 @@ export interface FancyListType extends FancyCategory {
     id: number;
     name: string;
     price: number;
-    tags: object;
+    tags: {
+      id: number;
+      name: string;
+      createdAt: string;
+    }[];
     image: string;
   }[];
 }
